fix(init): disable command on non-Linux platforms
