+++ cmd/init/filesystems.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package init
 
 import (
